+++ demo/Field/FieldDemo.jsx
@@ -75,6 +75,14 @@ class Demo extends React.Component {
             <div className="t-omit">自定义输入元素的区域多行点点点很多很多很多很多很多很多很多很多很多很多很多很多很多很多很多很多很多很多</div>
           </Field>
         </Group.List>
+        <Group.List>
+          <Field layout="v" label="上下结构（右侧自定义）" labelRight={<AngleRight {...angleIconProps} />}>
+            <div className="t-omit">自定义输入元素的区域</div>
+          </Field>
+        </Group.List>
+        <Group.List>
+          <Field layout="v" label="上下结构（没有 children）" tip={<div>这里是tip这里是tip这里是tip这里是tip这里是tip这里是tip这里是tip这里是tip<a href="https://salt-ui.github.io/" target="_blank">salt-ui</a>这里是tip</div>} />
+        </Group.List>
       </div>
     );
   }