+++ src/ScrollView/ScrollView.jsx
@@ -49,6 +49,10 @@ class ScrollView extends React.Component {
     }
   }
 
+  componentWillUnmount() {
+    this.onScroll.cancel();
+  }
+
   doScroll() {
     const { scrollTop } = this.scrollView;
     const onScroll = this.props.infiniteScrollOptions.onScroll || Context.noop;
@@ -62,6 +66,7 @@ class ScrollView extends React.Component {
     if (refreshControl) {
       element = (<RefreshControl
         key="refreshControl"
+        getScrollContainer={() => this.scrollView}
         {...refreshControlOptions}
         className={Context.prefixClass('scroll-view-inner')}
       />);
@@ -101,7 +106,7 @@ class ScrollView extends React.Component {
           {...this.props.infiniteScrollOptions}
           getDOMNode={(node) => {
             if (node) {
-              this.scrollNode = node;
+              this.scrollView = node;
             }
           }}
         >
@@ -110,7 +115,7 @@ class ScrollView extends React.Component {
       );
       this.hasInfiniteScroll = true;
     } else if (this.hasInfiniteScroll) {
-      this.scrollTop = this.scrollNode.scrollTop;
+      this.scrollTop = this.scrollView.scrollTop;
     }
 
     return element;